La correction de la modification d'un livre

+++ controllers/book_ctrl.js
@@ -92,13 +92,13 @@ exports.modifyBook = (req, res, next) => {
             if (book.userId != req.auth.userId) {
                 res.status(401).json({ message: 'Non-autorisé' });
             } else {
-                if (book.imageUrl) {
+                if (req.file && book.imageUrl) {
                     const oldImageName = book.imageUrl.split('/images/')[1];
                     fs.unlink(`images/${oldImageName}`, (err) => {
                         if (err) {
                             console.error('Ancienne image non suprimée', err);
                         }
-                    })
+                    });
                 }
                 Book.updateOne({ _id: req.params.id}, { ...bookObject, _id: req.params.id})
                 .then(() => res.status(200).json({message: 'Livre modifié!'}))
